feat: add ability to draw connections between nodes
This commit fixes a bug where the hidden canvas was not being redrawn when the window was resized. Additionally, it adds a new feature that allows connections to be drawn between nodes in the circuit.

The coordinate system for the canvas has been simplified to make it easier to work with. The gate classes have also been simplified to reduce complexity and repetition.

The logic has been refactored so some gates can handle an arbitrary number of inputs, allowing for more complex circuits to be created.